pyproject.toml
[project]
name = "photobooth"
version = "0.1.0"
description = ""
authors = [
    {name = "Adrian",email = "adrian.rosian@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "google-genai (>=1.19.0,<2.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "boto3 (>=1.38.34,<2.0.0)",
    "qrcode[pil] (>=8.2,<9.0)",
    "sounddevice (>=0.5.2,<0.6.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "aws-cdk-lib (>=2.200.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
